Moved Item component to own file from explorer
