+++ backend/App/client/src/pages/Explorer/Explorer.jsx
@@ -2,13 +2,13 @@ import React, { useRef } from "react";
 import { useEffect } from "react";
 import { useState } from "react";
 import { useContext } from "react";
-import { NavLink } from "react-router-dom";
 import { AuthContext } from "../../auth/AuthContext";
 import { Outlet, useNavigate, useParams } from "react-router";
 import Loader from "../../components/Loaders/Loader";
 import { CategoryContext } from "../../contexts/CategoryContext";
 import "./Explorer.scss";
 import ContextMenu from "./components/ContextMenu";
+import Item from "./ExplorerItem";
 
 const Explorer = () => {
   const categoryContext = useContext(CategoryContext);
@@ -151,75 +151,3 @@ const Explorer = () => {
 };
 
 export default Explorer;
-
-const Item = ({ category, recentPath, setCurrentContextItem }) => {
-  const [showChilds, setShowChilds] = useState(false);
-
-  useEffect(() => {
-    if (recentPath?.includes(category.categoryId)) {
-      setShowChilds(true);
-    }
-  }, []);
-
-  return (
-    <>
-      <NavLink
-        to={`/explore/${category.categoryId}`}
-        className="item ellipsis"
-        style={determineDepth(category)}
-        onContextMenu={(e) => {
-          e.stopPropagation();
-          setCurrentContextItem(category);
-        }}
-      >
-        {category.childCategories ? (
-          showChilds ? (
-            <i
-              onClick={(e) => {
-                e.preventDefault();
-                setShowChilds(!showChilds);
-              }}
-              class="fa fa-angle-down"
-            ></i>
-          ) : (
-            <i
-              onClick={(e) => {
-                e.preventDefault();
-                setShowChilds(!showChilds);
-              }}
-              class="fa fa-angle-right"
-            ></i>
-          )
-        ) : (
-          <i style={{ marginLeft: "10px" }}></i>
-        )}
-        {category.name}
-      </NavLink>
-      {showChilds ? (
-        category.childCategories?.map((subCategory) => {
-          return (
-            <Item
-              key={subCategory.categoryId}
-              category={subCategory}
-              recentPath={recentPath}
-              setCurrentContextItem={setCurrentContextItem}
-            />
-          );
-        })
-      ) : (
-        <></>
-      )}
-    </>
-  );
-};
-
-const determineDepth = (category) => {
-  if (category.level > 0) {
-    const style = {
-      marginLeft: `${category.level * 5 + 10}px`,
-      border: "none",
-    };
-
-    return style;
-  }
-};
