+++ backend/App/client/src/pages/Explorer/ExplorerItem.jsx
@@ -1,0 +1,77 @@
+import { useEffect } from "react";
+import { useState } from "react";
+import { NavLink } from "react-router-dom";
+
+const Item = ({ category, recentPath, setCurrentContextItem }) => {
+  const [showChilds, setShowChilds] = useState(false);
+
+  useEffect(() => {
+    if (recentPath?.includes(category.categoryId)) {
+      setShowChilds(true);
+    }
+  }, []);
+
+  return (
+    <>
+      <NavLink
+        to={`/explore/${category.categoryId}`}
+        className="item ellipsis"
+        style={determineDepth(category)}
+        onContextMenu={(e) => {
+          e.stopPropagation();
+          setCurrentContextItem(category);
+        }}
+      >
+        {category.childCategories ? (
+          showChilds ? (
+            <i
+              onClick={(e) => {
+                e.preventDefault();
+                setShowChilds(!showChilds);
+              }}
+              class="fa fa-angle-down"
+            ></i>
+          ) : (
+            <i
+              onClick={(e) => {
+                e.preventDefault();
+                setShowChilds(!showChilds);
+              }}
+              class="fa fa-angle-right"
+            ></i>
+          )
+        ) : (
+          <i style={{ marginLeft: "10px" }}></i>
+        )}
+        {category.name}
+      </NavLink>
+      {showChilds ? (
+        category.childCategories?.map((subCategory) => {
+          return (
+            <Item
+              key={subCategory.categoryId}
+              category={subCategory}
+              recentPath={recentPath}
+              setCurrentContextItem={setCurrentContextItem}
+            />
+          );
+        })
+      ) : (
+        <></>
+      )}
+    </>
+  );
+};
+
+const determineDepth = (category) => {
+  if (category.level > 0) {
+    const style = {
+      marginLeft: `${category.level * 5 + 10}px`,
+      border: "none",
+    };
+
+    return style;
+  }
+};
+
+export default Item;
